Make Resource and ResourceIdentifier siblings.
