+++ src/Resource.php
@@ -7,7 +7,7 @@ namespace silverorange\JsonApiClient;
 use silverorange\JsonApiClient\Exception\InvalidPropertyException;
 use silverorange\JsonApiClient\Exception\InvalidResourceTypeException;
 
-abstract class Resource extends ResourceIdentifier
+abstract class Resource extends AbstractResource
 {
     // {{{ class constants
